+++ src/mainComic.c
@@ -2,7 +2,6 @@
 #include <stdlib.h>
 
 #include "Comic.h"
-#include "PNM.h"
 
 
 /* ------------------------------------------------------------------------- *\
@@ -46,7 +45,7 @@ int main(int argc, char* argv[])
     }
 
 
-    size_t nbImages = argc - 4;
+    size_t nbImages = (size_t) (argc - 4);
 
     PNMImage** images = (PNMImage**) malloc((nbImages) * sizeof(PNMImage*));
     if (!images)
